Readme fixes
Made it actually English and standardized the period usage in instructions (half had periods, half didn't, so removing them entirely makes it look better ig, you can change it so all instructions have full stops if you want to). Also fixed an issue in IntelliJ instructions, it's Lifecycle not Lifestyle, and it's listed as Maven in View -> Tools not Maven Projects (if it says that it is maven projects then update your fucking IDE will). I'll make it sexy later just commit this for now.